courseDetails-función para traerse detalles de curso
